Fixing lint and build issues
