+++ frontend/packages/core/src/utils/index.ts
@@ -1,3 +1,2 @@
-// eslint-disable-next-line import/prefer-default-export
 export { default as getDisplayName } from "./getDisplayName";
 export { default as findPathMatchList } from "./pathMatching";
