+++ frontend/workflows/projectCatalog/src/helpers/link-text.tsx
@@ -1,6 +1,5 @@
 import React from "react";
 import { Link, Typography } from "@clutch-sh/core";
-import * as _ from "lodash";
 
 const LinkText = ({ text, link }: { text: string; link?: string }) => {
   const returnText = <Typography variant="body2">{text}</Typography>;
