+++ frontend/workflows/projectCatalog/src/types/config.ts
@@ -1,0 +1,5 @@
+export interface ProjectConfigLink {
+  title: string;
+  path: string;
+  icon?: React.ReactElement;
+}
