+++ frontend/workflows/projectCatalog/src/types/index.ts
@@ -1,0 +1,3 @@
+export * from "./config";
+export * from "./details";
+export * from "./workflow";
