+++ frontend/workflows/projectCatalog/src/types/workflow.ts
@@ -1,0 +1,7 @@
+import type { BaseWorkflowProps } from "@clutch-sh/core";
+
+export interface ProjectCatalogProps {
+  allowDisabled?: boolean;
+}
+
+export interface WorkflowProps extends BaseWorkflowProps, ProjectCatalogProps {}
